Rename Protocol to Bridge, move to own file
